Function riscv_decode::instruction_length

pub fn instruction_length(i: u16) -> usize

Return the length (in bytes) of an instruction given the low 16 bits of it.

The current spec reserves a bit pattern for instructions of length >= 192 bits, but for simplicity this function just returns 24 in that case. The largest instructions currently defined are 4 bytes so it will likely be a long time until this diffence matters.